nest_mask_detection/requirements.txt
# Core ML/Computer Vision
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.195
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0

# API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2

# Google Cloud
google-cloud-pubsub==2.18.4
google-cloud-storage==2.10.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Web requests
requests==2.31.0

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Development (optional)
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# Logging and monitoring
python-json-logger==2.0.7
